Added an overflow check to SID stream buffers (#620)

+++ Emulator/SID/ReSID.h
@@ -226,6 +226,6 @@ public:
      * samples are written into the provided ring buffer. The fuction returns
      * the number of written audio samples.
      */
-    i64 executeCycles(u64 numCycles, SampleStream &stream);
-    i64 executeCycles(u64 numCycles);
+    i64 executeCycles(usize numCycles, SampleStream &stream);
+    i64 executeCycles(usize numCycles);
 };
